% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{extractSpectrum}
\alias{extractSpectrum}
\title{Extract the spectrum}
\usage{
extractSpectrum(mtx, method = "ols", gate = T, powerGate = T, elim = NULL)
}
\description{
the spectra are scaled to the mI.
}
